move utilities in a folder to decect changes in watch mode
