+++ girder_vip/web_client/views/HeaderUserView.js
@@ -1,7 +1,7 @@
 // Import utilities
 import { wrap } from '@girder/core/utilities/PluginUtils';
 import { AccessType } from '@girder/core/constants';
-import { hasTheVipApiKeyConfigured } from '../utilities';
+import { hasTheVipApiKeyConfigured } from '../utilities/vipPluginUtils';
 import events from '@girder/core/events';
 
 // Import views
